+++ src/test/java/io/github/astrapi69/swing/base/BaseDialogExample.java
@@ -33,8 +33,8 @@ import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.IModel;
 import io.github.astrapi69.swing.button.ButtonFactory;
 import io.github.astrapi69.swing.layout.ScreenSizeExtensions;
-import io.github.astrapi69.swing.panels.login.pw.ChangePasswordModelBean;
-import io.github.astrapi69.swing.panels.login.pw.NewPasswordPanel;
+import io.github.astrapi69.swing.panel.login.pw.ChangePasswordModelBean;
+import io.github.astrapi69.swing.panel.login.pw.NewPasswordPanel;
 import io.github.astrapi69.window.adapter.CloseWindow;
 
 public class BaseDialogExample extends BaseDialog<ChangePasswordModelBean>